MAPREDUCE-7446. Fixed int overflow

+++ hadoop-mapreduce-project/hadoop-mapreduce-client/hadoop-mapreduce-client-core/src/main/java/org/apache/hadoop/mapred/IFile.java
@@ -434,7 +434,7 @@ public class IFile {
     }
     
     public void nextRawValue(DataInputBuffer value) throws IOException {
-      long targetSizeLong = currentValueLength + (currentValueLength >> 1);
+      long targetSizeLong = currentValueLength + (long) (currentValueLength >> 1);
       int targetSize = (int) Math.min(targetSizeLong, ARRAY_MAX_SIZE);
 
       final byte[] valBytes = (value.getData().length < currentValueLength)
